requirements.txt
pandas
numpy
scikit-learn
nltk
joblib
matplotlib
plotly
requests
transformers
torch
fastapi
uvicorn
pydantic
python-multipart
openai-whisper